pyproject.toml
[tool.poetry]
name = "{project-name}"
version = "0.1.0"
description = ""
authors = ["christopher-cirelli <ccirelli2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
python-decouple = "^3.8"
pre-commit = "^3.5.0"
pylint = "^3.0.2"
pytest = "^7.4.3"
gitpython = "^3.1.40"
jupyter = "^1.0.0"
jupyterlab = "^4.0.7"
pytest-cov = "^4.1.0"
streamlit = "^1.28.1"
black = "^23.11.0"
seaborn = "^0.13.0"
scikit-learn = "^1.3.2"
pydantic = "1.10.13"
huggingface-hub = "^0.31.1"
python-dotenv = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
